+++ src/be_/index.d.ts
@@ -1,13 +1,13 @@
 export declare const pending_symbol:unique symbol
 export declare function globalThis__be_<
-	Val extends NonNullable<unknown>,
+	Val,
 	ctx_T extends Ctx = Ctx
 >(
 	val_:be__val__T<Val, ctx_T>,
 	be__params?:be__params_T
 ):Be<Val, ctx_T>
 export declare function globalThis__be_<
-	Val extends NonNullable<unknown>,
+	Val,
 	ctx_T extends Ctx = Ctx
 >(
 	id:string,
@@ -21,14 +21,14 @@ export declare function globalThis__be_<
  * otherwise it creates the value using the val_ factory function.
  */
 export declare function be_<
-	Val extends NonNullable<unknown>,
+	Val,
 	ctx_T extends Ctx = Ctx
 >(
 	val_:be__val__T<Val, ctx_T>,
 	be__params?:be__params_T
 ):Be<Val, ctx_T>
 export declare function be_<
-	Val extends NonNullable<unknown>,
+	Val,
 	ctx_T extends Ctx = Ctx
 >(
 	id:string,
@@ -40,63 +40,76 @@ export {
 	be_ as b_,
 	be_ as _b,
 }
-export declare function be__is_source_(be:Be<any>):(map_ctx:MapCtx, ctx:Ctx)=>boolean
-export declare function source__map_ctx_(ctx:Ctx, is_source_:is_source__T):MapCtx
+export declare function be__is_source_<
+	ctx_T extends Ctx = Ctx
+>(be:Be<any>):(map_ctx:MapCtx, ctx:ctx_T)=>boolean
+export declare function source__map_ctx_<
+	ctx_T extends Ctx = Ctx
+>(ctx:Ctx, is_source_:is_source__T):MapCtx
 export declare function be__set<
-	Val extends NonNullable<unknown>,
+	Val,
 	ctx_T extends Ctx = Ctx
 >(
 	be:Be<Val>,
-	ctx:Ctx,
+	ctx:ctx_T,
 	val:Val
 ):void
 export declare function ctx__set<
-	Val extends NonNullable<unknown>,
+	Val,
 	ctx_T extends Ctx = Ctx
 >(
-	ctx:Ctx,
+	ctx:ctx_T,
 	be_OR_id:Be<Val>|string|symbol,
 	val:Val,
 	is_source_?:is_source__T
 ):void
 export declare function be__delete<
-	Val extends NonNullable<unknown>
+	Val,
+	ctx_T extends Ctx = Ctx
 >(
 	be:Be<Val>,
-	ctx:Ctx
+	ctx:ctx_T
 ):void
 export declare function ctx__delete<
-	Val extends NonNullable<unknown>
+	Val,
+	ctx_T extends Ctx = Ctx
 >(
-	ctx:Ctx,
+	ctx:ctx_T,
 	be_OR_id:Be<Val>|string|symbol,
 	is_source_?:is_source__T
 ):void
+export declare function be__has_<
+	ctx_T extends Ctx = Ctx
+>(be_OR_id:Be<any>|string, ctx:ctx_T):boolean
+export declare function be__has__ctx_<
+	ctx_T extends Ctx = Ctx
+>(be_OR_id:Be<any>|string, ctx:ctx_T):MapCtx
 export declare function be__val_<
-	Val extends NonNullable<unknown>
->(be_OR_id:Be<Val>|string, ctx:Ctx):Val|unknown|null
+	Val,
+	ctx_T extends Ctx = Ctx
+>(be_OR_id:Be<Val>|string, ctx:ctx_T):Val|unknown|null
 export declare type MapCtx = Map<Be<any>|string|symbol, unknown>
 export interface NestedMapCtx extends Array<NestedMapCtx|MapCtx> {
 }
 export type Ctx = MapCtx|NestedMapCtx
 export declare type Be<
-	Val extends NonNullable<unknown>,
+	Val,
 	ctx_T extends Ctx = Ctx
 > = ((ctx:ctx_T, params?:be_params_T)=>Val)&{ id?:string }
 export declare type B<
-	Val extends NonNullable<unknown>,
+	Val,
 	ctx_T extends Ctx = Ctx
 > = Be<Val, ctx_T>
 export declare type be__return_T<
-	Val extends NonNullable<unknown>,
+	Val,
 	ctx_T extends Ctx = Ctx
 > = Be<Val, ctx_T>
 export declare type be_T<
-	Val extends NonNullable<unknown>,
+	Val,
 	ctx_T extends Ctx = Ctx
 > = Be<Val, ctx_T>
 export declare type be__val__T<
-	Val extends NonNullable<unknown>,
+	Val,
 	ctx_T extends Ctx = Ctx
 > = (ctx:ctx_T, be:Be<Val, ctx_T>, params?:be_params_T)=>Val
 export interface be__params_T {