+++ src/be_/index.js
@@ -1,7 +1,10 @@
 import { isArray } from '@ctx-core/function'
 import { globalThis__prop__ensure } from '../globalThis__prop__ensure/index.js'
-const pending_sym = Symbol.for('pending')
-const be_M_is_source_ = globalThis__prop__ensure(
+let proto_ = Object.getPrototypeOf
+let string_proto = proto_('')
+let _null = null
+let pending_sym = Symbol.for('pending')
+let be_M_is_source_ = globalThis__prop__ensure(
 	Symbol.for('be_M_is_source_'),
 	()=>new WeakMap())
 /** @typedef {import('./index.d.ts').Be}Be */
@@ -37,7 +40,7 @@ export function globalThis__be_(
 /**
  * Auto-memoization function for the Ctx.
  *
- * Returns a function to ensure that a member id is defined on a ctx object,
+ * Returns a function to ensure that a member id is defined on a Ctx,
  * otherwise it creates the value using the val_ factory function.
  * @param {string|be__val__T}id_OR_val__new
  * @param {be__val__T|be__params_T}[val__new_OR_be__params]
@@ -48,43 +51,43 @@ export function globalThis__be_(
 export function be_(
 	id_OR_val__new,
 	val__new_OR_be__params,
-	be__params
+	be__params = {}
 ) {
 	/** @type {string} */
-	const id =
-		typeof id_OR_val__new === 'string'
-		? id_OR_val__new
-		: null
+	let id =
+		proto_(id_OR_val__new) === string_proto
+			? id_OR_val__new
+			: _null
 	/** @type {be__val__T} */
-	const val__new =
-		typeof id_OR_val__new === 'string'
-		? val__new_OR_be__params
-		: id_OR_val__new
+	let val__new =
+		proto_(id_OR_val__new) === string_proto
+			? val__new_OR_be__params || (()=>null)
+			: id_OR_val__new
 	be__params =
-		typeof id_OR_val__new === 'string'
-		? be__params
-		: val__new_OR_be__params
-	const is_source_ =
+		proto_(id_OR_val__new) === string_proto
+			? be__params
+			: val__new_OR_be__params
+	let is_source_ =
 		be__params
-		? be__params.is_source_
-		: null
-	const expired_ =
+			? be__params.is_source_
+			: _null
+	let expired_ =
 		be__params
-		? be__params.expired_
-		: null
-	const be = (argv__ctx, params)=>{
+			? be__params.expired_
+			: ()=>0
+	let be = (argv__ctx, params = {})=>{
 		if (!argv__ctx) {
 			throw new Error(`be must have a Ctx passed as an argument`)
 		}
-		const saved__val = be__val_(be, argv__ctx)
+		let saved__val = be__val_(be, argv__ctx)
 		if (
-			saved__val !== undefined
-			&& (!params || !params.force)
+			be__has_(be, argv__ctx)
+			&& !params.force
 			&& (!expired_ || !expired_(argv__ctx))
 		) {
 			return saved__val
 		}
-		const ctx = source__map_ctx_(argv__ctx, is_source_)
+		let ctx = source__map_ctx_(argv__ctx, is_source_)
 		if (!ctx) {
 			throw new Error(
 				`be: ${
@@ -97,8 +100,8 @@ export function be_(
 			ctx.set(pending_sym, pending)
 		}
 		if (pending.get(be)) {
-			const pending_value_a = []
-			for (const value of pending.values()) {
+			let pending_value_a = []
+			for (let value of pending.values()) {
 				pending_value_a.push(value)
 			}
 			throw new Error(
@@ -106,26 +109,12 @@ export function be_(
 					String(id)
 				}: circular dependency:\n${pending_value_a.map(pending_value=>
 					typeof pending_value === 'function'
-					? 'Function'
-					: pending_value).join('\n')}`)
+						? 'Function'
+						: pending_value).join('\n')}`)
 		}
 		pending.set(be, id || be)
-		const val = val__new(argv__ctx, be, params)
-		if (ctx.get(be) === undefined) {
-			if (val === undefined) {
-				throw new Error(
-					`be_: ${
-						id
-						? String(id) + ': '
-						: ''
-					}: val_ must return a non-undefined value. ${
-						id
-						? ''
-						: '\n' + val__new.toString()
-					}`)
-			}
-			ctx.set(be, val)
-		}
+		let val = val__new(argv__ctx, be, params)
+		ctx.set(be, val)
 		if (id) {
 			ctx.set(id, val)
 		}
@@ -146,7 +135,7 @@ export {
  * @returns {is_source__T}
  * @private
  */
-export function be__is_source_(be) {
+export function be__is_source__(be) {
 	return be_M_is_source_.get(be)
 }
 /**
@@ -157,7 +146,7 @@ export function be__is_source_(be) {
  * @private
  */
 export function be__set(be, ctx, val) {
-	ctx__set(ctx, be, val, be__is_source_(be))
+	ctx__set(ctx, be, val, be__is_source__(be))
 }
 /**
  * @param {Ctx}ctx
@@ -173,13 +162,14 @@ export function ctx__set(
 	val,
 	is_source_
 ) {
-	if (is_source_ == null) is_source_ = ()=>true
-	const source__map_ctx = source__map_ctx_(ctx, is_source_)
-	if (!source__map_ctx) return
-	source__map_ctx.set(be_OR_id, val)
-	const { id } = be_OR_id
-	if (id) {
-		source__map_ctx.set(id, val)
+	if (!is_source_) is_source_ = ()=>true
+	let source__map_ctx = source__map_ctx_(ctx, is_source_)
+	if (source__map_ctx) {
+		source__map_ctx.set(be_OR_id, val)
+		let { id } = be_OR_id
+		if (id) {
+			source__map_ctx.set(id, val)
+		}
 	}
 }
 /**
@@ -188,7 +178,7 @@ export function ctx__set(
  * @private
  */
 export function be__delete(be, ctx) {
-	ctx__delete(ctx, be, be__is_source_(be))
+	ctx__delete(ctx, be, be__is_source__(be))
 }
 /**
  * @param {Ctx}ctx
@@ -200,7 +190,7 @@ export function ctx__delete(
 	be_OR_id,
 	is_source_
 ) {
-	if (is_source_ == null) is_source_ = ()=>true
+	if (!is_source_) is_source_ = ()=>true
 	if (isArray(ctx)) {
 		for (let i = 0; i < ctx.length; i++) {
 			if (is_source_(ctx[i], ctx)) {
@@ -209,14 +199,40 @@ export function ctx__delete(
 		}
 	} else {
 		/** @type {MapCtx} */
-		const map_ctx = /** @type {any} */ctx
+		let map_ctx = /** @type {any} */ctx
 		if (is_source_(map_ctx, ctx)) {
-			const { id } = be_OR_id
+			let { id } = be_OR_id
 			if (id) {
 				map_ctx.delete(id)
 			}
 			map_ctx.delete(be_OR_id)
 		}
+	}
+}
+/**
+ * @param {Be|string}be_or_id
+ * @param {Ctx}argv__ctx
+ * @returns {boolean}
+ * @private
+ */
+export function be__has_(be_or_id, argv__ctx) {
+	return !!be__has__ctx_(be_or_id, argv__ctx)
+}
+/**
+ * @param {Be|string}be_or_id
+ * @param {Ctx}argv__ctx
+ * @returns {MapCtx}
+ * @private
+ */
+export function be__has__ctx_(be_or_id, argv__ctx) {
+	if (isArray(argv__ctx)) {
+		for (let i = 0; i < argv__ctx.length; i++) {
+			const be__has__ctx = be__has__ctx_(be_or_id, argv__ctx[i])
+			if (be__has__ctx) return be__has__ctx
+		}
+		return null
+	} else {
+		return argv__ctx.has(be_or_id) ? argv__ctx : null
 	}
 }
 /**
@@ -228,14 +244,12 @@ export function ctx__delete(
 export function be__val_(be_or_id, argv__ctx) {
 	if (isArray(argv__ctx)) {
 		for (let i = 0; i < argv__ctx.length; i++) {
-			const ctx = argv__ctx[i]
-			const saved__val = be__val_(be_or_id, ctx)
-			if (saved__val !== undefined) return saved__val
+			let ctx = argv__ctx[i]
+			const be__has__ctx = be__has__ctx_(be_or_id, ctx)
+			if (be__has__ctx) return be__has__ctx.get(be_or_id)
 		}
 	} else {
-		/** @type {MapCtx} */
-		const map_ctx = /** @type {any} */argv__ctx
-		return map_ctx.get(be_or_id)
+		return argv__ctx.get(be_or_id)
 	}
 }
 /**
@@ -247,9 +261,9 @@ export function be__val_(be_or_id, argv__ctx) {
 export function source__map_ctx_(ctx, is_source_) {
 	if (isArray(ctx)) {
 		for (let i = 0; i < ctx.length; i++) {
-			const i_ctx = ctx[i]
-			const first__ctx = source__map_ctx_(i_ctx, is_source_)
-			if (first__ctx) return first__ctx
+			let i_ctx = ctx[i]
+			let source__map_ctx = source__map_ctx_(i_ctx, is_source_)
+			if (source__map_ctx) return source__map_ctx
 		}
 	} else if (!is_source_ || is_source_(/** @type {MapCtx} */ctx, ctx)) {
 		return ctx